requirements.txt
Django>=4.2
opencv-python>=4.5
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
